config: rename VelRenderTextureProject project
