+++ VelaptorTemplates.sln
@@ -1,6 +1,5 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-#
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VelaptorTemplates", "VelaptorTemplates\VelaptorTemplates.csproj", "{D85354EC-A891-44B1-829D-EDDD592693DF}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "VelaptorTemplates", "VelaptorTemplates", "{86B02555-09AE-4020-A13E-391E432A2821}"
@@ -9,7 +8,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Templates", "Templates", "{
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VelEmpty", "VelaptorTemplates\Templates\VelEmpty\VelEmpty.csproj", "{2AA3B934-D685-4C45-9F80-4B5C0AF96CB5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VelRenderTexture", "VelaptorTemplates\Templates\VelRenderTexture\VelRenderTexture.csproj", "{2F8A49F7-1DF4-4D15-9649-6111ABD91BF4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VeleptorRenderTexture", "VelaptorTemplates\Templates\VelaptorRenderTexture\VeleptorRenderTexture.csproj", "{2F8A49F7-1DF4-4D15-9649-6111ABD91BF4}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionItems", "SolutionItems", "{1A2DFA7D-1B2B-4CF4-A292-15419F2579C7}"
 	ProjectSection(SolutionItems) = preProject
@@ -39,6 +38,9 @@ Global
 		{2F8A49F7-1DF4-4D15-9649-6111ABD91BF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2F8A49F7-1DF4-4D15-9649-6111ABD91BF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2F8A49F7-1DF4-4D15-9649-6111ABD91BF4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{D85354EC-A891-44B1-829D-EDDD592693DF} = {86B02555-09AE-4020-A13E-391E432A2821}
