+++ VelaptorTemplates/Templates/VelaptorRenderTexture/Program.cs
@@ -1,4 +1,4 @@
-﻿using VelRenderTexture;
+﻿using VelaptorRenderTexture;
 
 // Create a new game instance and run it to start the game
 var game = new Game();
